Tweaked constants in memory model.
The main motivation and difference is that you now get nicer, round numbers
displayed in the UI, progressing roughly from 0 to 100% in steps of 5% each
time if you ace the tests, rather than an initial jump from 0 to 12%,
followed by slower progress.

Due to the small changes in the constants, there is no need to adjust
existing 'strength' values for the new algorithm.